fastapi-growth-ai/requirements.txt
absl-py==2.0.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
auth0-python @ git+https://github.com/auth0/auth0-python.git@18e4defa3efe7d12d83ffe090b9f3fd3998c9045
bcrypt==4.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
databases==0.8.0
docopt==0.6.2
ecdsa==0.18.0
email-validator==2.0.0.post2
environs==9.5.0
fastapi==0.103.1
fastapi-auth==0.1.1
fastapi-auth0==0.5.0
flatbuffers==23.5.26
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.59.0
h11==0.14.0
httptools==0.6.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
libclang==16.0.6
makefun==1.15.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
multidict==6.0.4
mysql-connector-python==8.1.0
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
orjson==3.9.7
packaging==23.1
pandas==2.1.1
passlib==1.7.4
Pillow==10.1.0
pipdeptree==2.13.0
pipreqs==0.4.13
protobuf==4.21.12
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.27.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.5
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
Werkzeug==3.0.0
wrapt==1.14.1
yarg==0.1.9
yarl==1.9.2
